Move driver to new class
